pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "oculimundi"
version = "0.1.0"
description = "Oculi Mundi - prowizoryczna strzelanka napisana przy pomocy biblioteki PyGame"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=2.0.1",
    "pygame>=2.5.2",
    "PyOpenGL>=3.1.7",
    "PyOpenGL-accelerate>=3.1.7"
]

[project.urls]
source = "https://github.com/Barionetta/oculi-mundi"

[project.scripts]
oculimundi = "oculimundi:main"